+++ app/src/main/java/com/example/help_code/presentation/blank/CoroutineViewModel.kt
@@ -1,7 +1,56 @@
 package com.example.help_code.presentation.blank
 
-import com.example.help_code.base.CompositeViewModel
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.example.help_code.data.testgateway.CMSGateway
+import kotlinx.coroutines.*
+import retrofit2.HttpException
+import timber.log.Timber
 
-class CoroutineViewModel:CompositeViewModel() {
+val myTimber = Timber.tag("BlankViewModel")
 
+val listUrl = listOf(
+    "https://example1.com",
+    "https://testApi.com/test",
+    "https://rdsig.yahoo.co.jp/smptop/topi/test",
+    "https://weu-api-digital-tst.azure-api.net/api/member-details/v2/updates/30218711",
+    "https://weu-api-digital-tst.azure-api.net/api/member-details/v2/updates",
+    "https://weu-api-digital-tst.azure-api.net/api/member-details/v2",
+    "https://klassbook.commonsware.com"
+)
+
+class BlankViewModel : ViewModel() {
+
+    private val gateway = CMSGateway()
+
+    val handleError = CoroutineExceptionHandler { _, error ->
+        myTimber.e("CoroutineExceptionHandler")
+        myTimber.e(error)
+    }
+
+    fun runTask() {
+        myTimber.w("coroutine above")
+        viewModelScope.launch(handleError) {
+            val result = mergeArray()
+            myTimber.w("Finally result: \n${result.joinToString(separator = "\n")}")
+        }
+        myTimber.w("coroutine below")
+    }
+
+    private suspend fun mergeArray() = supervisorScope {
+        listUrl.map { url ->
+            async {
+                val urltemp = url
+                kotlin.runCatching {
+                    gateway.getCMSCall(url)
+                }.onFailure {
+                    val _url = if (it is HttpException) it.response()
+                        ?.raw()?.networkResponse?.request?.url.toString() else urltemp
+                    myTimber.e("failure: ${it.message}, url: ${_url}")
+                }.onSuccess {
+                    myTimber.d("success: $it")
+                }
+            }
+        }.awaitAll()
+    }
 }
